pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lifeops"
version = "0.2.0"
requires-python = ">=3.11"

dependencies = [
    # LangGraph + LangChain
    "langgraph>=0.3",
    "langchain-core>=0.3",
    "langchain-groq>=0.2",
    "langchain-openai>=0.2",
    "langchain-mcp-adapters>=0.1",

    # MCP server
    "mcp[cli]>=1.0",
    "fastmcp>=0.1",

    # API layer
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",

    # Persistence
    "langgraph-checkpoint-postgres>=2.0",
    "psycopg[binary,pool]>=3.2",
    "asyncpg>=0.29",

    # HTTP client (used in MCP servers)
    "httpx>=0.27",

    # Scheduling
    "apscheduler>=3.10",
    "sqlalchemy>=2.0",

    # Web search
    "tavily-python>=0.5",

    # NLP / date parsing
    "dateparser>=1.2",

    # Settings
    "pydantic-settings>=2.0",
    "pydantic>=2.0",

    # Auth
    "bcrypt>=4.0",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx"]

[tool.hatch.build.targets.wheel]
packages = ["src"]